refactor(gpu): int32 dense count arrays for the row-split allcounts path
Per-candidate counts are bounded by n_transactions, which the row-split
entry guards to < 2^31, and the <=8-GPU sum of per-shard partials is
bounded by that same n_transactions — so int64 dense count arrays were
unexamined legacy, not a requirement. count_k3plus_dense and
count_pairs_k2_dense now emit int32 and their wrappers allocate int32,
halving the dominant per-chunk VRAM term and the NCCL reduce traffic.
Hosts already widen to int64 right after threshold filtering.

+++ src/et_miner/gpu/kernels/_src/k3plus_dense.cu
@@ -10,7 +10,7 @@ void count_k3plus_dense(
     const long long n_u64s,
     const long long n_groups,  // int64: >2.1B groups at K>=9
     const long long total_candidates,
-    long long* __restrict__ result_counts,
+    int* __restrict__ result_counts,  // int32: counts <= n_transactions < 2^31 (guarded host-side)
     const long long candidate_offset
 ) {
     long long cand_idx = (long long)blockIdx.y * (long long)gridDim.x
@@ -74,6 +74,6 @@ void count_k3plus_dense(
         unsigned long long total = 0;
         int n_warps = (blockDim.x + 31) / 32;
         for (int w = 0; w < n_warps; w++) total += warp_sums[w];
-        result_counts[cand_idx - candidate_offset] = (long long)total;
+        result_counts[cand_idx - candidate_offset] = (int)total;
     }
 }
